+++ wire/cli/src/cli.rs
@@ -4,6 +4,7 @@ use clap_num::number_range;
 use clap_verbosity_flag::WarnLevel;
 use lib::SubCommandModifiers;
 use lib::hive::node::{Goal as HiveGoal, Name, SwitchToConfigurationGoal};
+use std::io::IsTerminal;
 
 use std::{
     fmt::{self, Display, Formatter},
@@ -28,8 +29,8 @@ pub struct Cli {
     #[arg(long, global = true, default_value = std::env::current_dir().unwrap().into_os_string())]
     pub path: std::path::PathBuf,
 
-    /// Hide progress bars
-    #[arg(long, global = true, default_value_t = false)]
+    /// Hide progress bars. Defaults to true if stdin does not refer to a tty (unix pipelines, in CI).
+    #[arg(long, global = true, default_value_t = !std::io::stdin().is_terminal())]
     pub no_progress: bool,
 
     /// Show trace logs