+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Wire now compiles and includes key agents for multiple architectures, currently only linux.
 - There is a new package output, `wire-small`, for testing purposes.
   It only compiles the key agent for the host that builds `wire-small`.
+- `--no-progress` now defaults to true if stdin does not refer to a tty (unix pipelines, in CI).
 
 ## [0.2.0] - 2025-04-21